SDK-100: resolved static analysis logic error of type:
"Comparing a pointer value of type 'NSNumber *' to a scalar integer value; did you mean to compare the result of calling a method on 'NSNumber *' to get the scalar value?"

+++ CleverTapSDK/ABTesting/uieditor/transformers/CTNSDictionaryFromUIImageValueTransformer.m
@@ -24,7 +24,7 @@
     NSDictionary *imageDictionary = imagesArray[0];
     
     NSNumber *scale = imageDictionary[@"scale"];
-    if (scale <= 0) {
+    if (scale.floatValue <= 0) {
         scale = @(1);
     }
     if (imageDictionary[@"url"]) {
